+++ src/components/ui/toaster.tsx
@@ -5,8 +5,8 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/components/ui/toast"
-import { useToast } from "@/components/ui/use-toast"
+} from "@/components/ui/Toast"
+import { useToast } from "@/hooks/useToast"
 
 export function Toaster() {
   const { toasts } = useToast()
